requirements.txt
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
exceptiongroup==1.2.2
h11==0.14.0
HumanCursor==1.1.5
idna==3.10
keyboard==0.13.5
MouseInfo==0.1.3
numpy==2.2.4
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pymongo==4.11.3
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.1
requests==2.32.3
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
undetected-chromedriver==3.5.5
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
